Linewidth now a double

+++ base/h/graphics.h
@@ -239,7 +239,7 @@ typedef struct _wcontext {
   struct SharedColor *fg, *bg;
   struct SharedPicture  *pattern;
   stringint     *linestyle;
-  int		linewidth;
+  double	linewidth;
   stringint     *drawop;
 #elif MSWIN32
   LOGPEN	pen;
